+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 - Simplified the syntax to modify registers. Use `mf.registers.arraystretch = 1.5`
   instead of mf.registers = {...mf.registers, arraystretch: 1.5}`
-
+- Allow changing registers using `\renewcommand`, for example
+  `\renewcommand{\arraystretch}{1.5}`
 
 ## 0.96.0 (2023-11-14)